Refactor the rest of implementation using DefaultRepository
