+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/extension/ResourceExtension.kt
@@ -547,12 +547,12 @@ suspend fun Task.updateDependentTaskDueDate(
  * to be a boolean otherwise the [toBoolean] function will evaluate to false and hence return an
  * empty list.
  */
-fun List<RepositoryResourceData>.filterByFhirPathExpression(
+fun Collection<RepositoryResourceData>.filterByFhirPathExpression(
   fhirPathDataExtractor: FhirPathDataExtractor,
   conditionalFhirPathExpressions: List<String>?,
   matchAll: Boolean,
 ): List<RepositoryResourceData> {
-  if (conditionalFhirPathExpressions.isNullOrEmpty()) return this
+  if (conditionalFhirPathExpressions.isNullOrEmpty()) return this as List
   return this.filter { repositoryResourceData ->
     if (matchAll) {
       conditionalFhirPathExpressions.all {
