+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/geowidget/GeoWidgetLauncherViewModel.kt
@@ -32,17 +32,14 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withContext
 import kotlinx.serialization.json.JsonPrimitive
 import org.hl7.fhir.r4.model.Enumerations
 import org.hl7.fhir.r4.model.Location
-import org.hl7.fhir.r4.model.ResourceType
 import org.smartregister.fhircore.engine.configuration.ConfigType
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.QuestionnaireConfig
 import org.smartregister.fhircore.engine.configuration.app.ApplicationConfiguration
 import org.smartregister.fhircore.engine.configuration.geowidget.GeoWidgetConfiguration
-import org.smartregister.fhircore.engine.configuration.register.ActiveResourceFilterConfig
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.domain.model.ActionParameter
 import org.smartregister.fhircore.engine.domain.model.ActionParameterType
@@ -101,124 +98,94 @@ constructor(
     searchText: String?,
   ) {
     viewModelScope.launch {
-      val totalCount =
-        withContext(dispatcherProvider.io()) {
-          defaultRepository.countResources(
-            filterByRelatedEntityLocation =
-              geoWidgetConfig.filterDataByRelatedEntityLocation == true,
-            baseResourceConfig = geoWidgetConfig.resourceConfig.baseResource,
-            filterActiveResources =
-              listOf(
-                ActiveResourceFilterConfig(
-                  resourceType = ResourceType.Patient,
-                  active = true,
-                ),
-                ActiveResourceFilterConfig(
-                  resourceType = ResourceType.Group,
-                  active = true,
-                ),
-              ),
+      val (locationsWithCoordinates, locationsWithoutCoordinates) =
+        defaultRepository
+          .searchNestedResources(
+            baseResourceIds = null,
+            fhirResourceConfig = geoWidgetConfig.resourceConfig,
             configComputedRuleValues = emptyMap(),
+            activeResourceFilters = null,
+            filterByRelatedEntityLocationMetaTag =
+              geoWidgetConfig.filterDataByRelatedEntityLocation == true,
+            currentPage = null,
+            pageSize = null,
           )
-        }
-      if (totalCount == 0L) {
-        showNoLocationDialog(geoWidgetConfig)
-        return@launch
-      }
-      var count = 0
-      var pageNumber = 0
-      var locationsWithoutCoordinatesCount = 0L
-      var registerDataCount = 0L
-      while (count < totalCount) {
-        val (locationsWithCoordinates, locationsWithoutCoordinates) =
-          defaultRepository
-            .searchResourcesRecursively(
-              filterActiveResources = null,
-              fhirResourceConfig = geoWidgetConfig.resourceConfig,
-              configRules = null,
-              secondaryResourceConfigs = null,
-              filterByRelatedEntityLocationMetaTag =
-                geoWidgetConfig.filterDataByRelatedEntityLocation == true,
-              currentPage = pageNumber,
-              pageSize = DefaultRepository.DEFAULT_BATCH_SIZE,
-            )
-            .asSequence()
-            .filter { it.resource is Location }
-            .partition {
-              with((it.resource as Location).position) { hasLongitude() && hasLatitude() }
-            }
-
-        val rules =
-          rulesExecutor.rulesFactory.generateRules(
-            geoWidgetConfig.servicePointConfig?.rules ?: emptyList(),
-          )
-
-        val registerData =
-          locationsWithCoordinates
-            .asSequence()
-            .map {
-              Pair(
-                it.resource as Location,
-                rulesExecutor.processResourceData(
-                  repositoryResourceData = it,
-                  rules = rules,
-                  params = emptyMap(),
-                ),
-              )
-            }
-            .map { (location, resourceData) ->
-              GeoJsonFeature(
-                id = location.logicalId,
-                geometry =
-                  Geometry(
-                    coordinates = // MapBox coordinates are represented as Long,Lat (NOT Lat,Long)
-                    listOf(
-                        location.position.longitude.toDouble(),
-                        location.position.latitude.toDouble(),
-                      ),
-                  ),
-                properties =
-                  geoWidgetConfig.servicePointConfig?.servicePointProperties?.mapValues {
-                    JsonPrimitive(it.value.interpolate(resourceData.computedValuesMap))
-                  } ?: emptyMap(),
-              )
-            }
-            .toList()
-        val features =
-          if (searchText.isNullOrBlank()) {
-            registerData
-          } else {
-            registerData.filter { geoJsonFeature: GeoJsonFeature ->
-              geoWidgetConfig.topScreenSection?.searchBar?.computedRules?.any { ruleName ->
-                // if ruleName not found in map return {-1}; check always return false hence no data
-                val value = geoJsonFeature.properties[ruleName]?.toString() ?: "{-1}"
-                value.contains(other = searchText, ignoreCase = true)
-              } == true
-            }
+          .values
+          .asSequence()
+          .filter { it.resource is Location }
+          .partition {
+            with((it.resource as Location).position) { hasLongitude() && hasLatitude() }
           }
 
-        geoJsonFeatures.postValue(features)
-
-        Timber.w(
-          locationsWithoutCoordinates.joinToString("\n") {
-            val position = (it.resource as Location).position
-            "Location id ${it.resource.logicalId} coordinates (${position.longitude},${position.latitude}) invalid."
-          },
+      val rules =
+        rulesExecutor.rulesFactory.generateRules(
+          geoWidgetConfig.servicePointConfig?.rules ?: emptyList(),
         )
-        pageNumber++
-        count += DefaultRepository.DEFAULT_BATCH_SIZE
-        registerDataCount += features.size
-        locationsWithoutCoordinatesCount += locationsWithoutCoordinates.size
-      }
 
-      val locationsCount = if (searchText.isNullOrBlank()) totalCount else registerDataCount
+      val registerData =
+        locationsWithCoordinates
+          .asSequence()
+          .map {
+            Pair(
+              it.resource as Location,
+              rulesExecutor.processResourceData(
+                repositoryResourceData = it,
+                rules = rules,
+                params = emptyMap(),
+              ),
+            )
+          }
+          .map { (location, resourceData) ->
+            GeoJsonFeature(
+              id = location.logicalId,
+              geometry =
+                Geometry(
+                  coordinates = // MapBox coordinates are represented as Long,Lat (NOT Lat,Long)
+                  listOf(
+                      location.position.longitude.toDouble(),
+                      location.position.latitude.toDouble(),
+                    ),
+                ),
+              properties =
+                geoWidgetConfig.servicePointConfig?.servicePointProperties?.mapValues {
+                  JsonPrimitive(it.value.interpolate(resourceData.computedValuesMap))
+                } ?: emptyMap(),
+            )
+          }
+          .toList()
+      val features =
+        if (searchText.isNullOrBlank()) {
+          registerData
+        } else {
+          registerData.filter { geoJsonFeature: GeoJsonFeature ->
+            geoWidgetConfig.topScreenSection?.searchBar?.computedRules?.any { ruleName ->
+              // if ruleName not found in map return {-1}; check always return false hence no data
+              val value = geoJsonFeature.properties[ruleName]?.toString() ?: "{-1}"
+              value.contains(other = searchText, ignoreCase = true)
+            } == true
+          }
+        }
+
+      geoJsonFeatures.postValue(features)
+
+      Timber.w(
+        locationsWithoutCoordinates.joinToString("\n") {
+          val position = (it.resource as Location).position
+          "Location id ${it.resource.logicalId} coordinates (${position.longitude},${position.latitude}) invalid."
+        },
+      )
+
+      val locationsCount =
+        if (searchText.isNullOrBlank()) {
+          locationsWithCoordinates.size + locationsWithoutCoordinates.size
+        } else features.size
 
       // Account for locations without coordinates
-      if (locationsWithoutCoordinatesCount in 1..locationsCount) {
+      if (locationsWithoutCoordinates.size in 1..locationsCount) {
         val message =
           context.getString(
             R.string.locations_without_coordinates,
-            locationsWithoutCoordinatesCount,
+            locationsWithoutCoordinates.size,
             locationsCount,
           )
         Timber.w(message)
@@ -246,7 +213,7 @@ constructor(
       }
 
       // Account for missing locations
-      if (locationsCount == 0L) {
+      if (locationsCount == 0) {
         if (!searchText.isNullOrBlank()) {
           val message =
             context.getString(
